requirements.txt
torch
torchvision
kornia
datasets
python-dotenv
matplotlib
scikit-image
